split stream and smarthome apis into different locations, added requirements.txt
