+++ api/requirements.txt
@@ -1,0 +1,3 @@
+fastapi==0.115.6
+opencv_python==4.10.0.84
+pyvesync==2.1.15
